continued migration from nephila ovs to endpoint
