+++ nephila/disabled.go
@@ -1,11 +1,5 @@
 package nephila
 
-import (
-	"fmt"
-
-	"github.com/Azure/azure-container-networking/netlink"
-)
-
 // DisabledNephilaProvider is returned when the provider is set to "Disabled"
 type DisabledNephilaProvider struct{}
 
@@ -13,17 +7,9 @@ func (fnp DisabledNephilaProvider) GetType() string {
 	return Disabled
 }
 
-func (dnp DisabledNephilaProvider) AddNetworkContainerRules(ovs NephilaOVSEndpoint, ncConfig interface{}) error {
-	return nil
-}
-func (dnp DisabledNephilaProvider) DeleteNetworkContainerRules(ovs NephilaOVSEndpoint, ncConfig interface{}) error {
-	return nil
-}
-func (dnp DisabledNephilaProvider) ConfigureNetworkContainerLink(link *netlink.VEthLink, ncConfig NephilaNetworkContainerConfig) error {
-	fmt.Printf("HERE DISABLED\n")
-	return nil
-}
 func (dnp DisabledNephilaProvider) ConfigureNode(nodeConf NephilaNodeConfig, dncConf NephilaDNCConfig) (NephilaNodeConfig, error) {
-	var nodeConfig NephilaNodeConfig
+	nodeConfig := NephilaNodeConfig{
+		Type: Disabled,
+	}
 	return nodeConfig, nil
 }
