+++ nephila/flannel.go
@@ -1,21 +1,37 @@
 package nephila
 
 import (
+	"bufio"
+	"context"
+	"encoding/json"
+	"errors"
+	"fmt"
 	"log"
 	"net"
+	"os"
 	"strconv"
+	"strings"
+	"time"
 
 	"github.com/Azure/azure-container-networking/netlink"
-	"github.com/Azure/azure-container-networking/ovsctl"
+	"github.com/coreos/etcd/client"
+	"github.com/coreos/flannel/pkg/ip"
 )
 
 const (
-	// Flannel type
-	Flannel = "Flannel"
-
-	// Disabled type
-	Disabled = "Disabled"
+	flannelKeyPath = "/coreos.com/network/config"
+	vxlan          = "vxlan"
 )
+
+type flannelEtcdBackendConfig struct {
+	Type string
+}
+
+type flannelEtcdConfig struct {
+	Network   ip.IP4Net
+	SubnetLen uint
+	Backend   flannelEtcdBackendConfig
+}
 
 // IPSubnet contains ip subnet.
 type IPSubnet struct {
@@ -47,46 +63,6 @@ func (fnp FlannelNephilaProvider) GetType() string {
 	return Flannel
 }
 
-func (fnp FlannelNephilaProvider) AddNetworkContainerRules(ovs NephilaOVSEndpoint, ncConfig interface{}) error {
-	log.Printf("HERE ADD RULES\n")
-	fncc := ncConfig.(NephilaNetworkContainerConfig)
-	config := fncc.Config.(FlannelNetworkContainerConfig)
-	nodeConfig := fncc.NodeConfig.(FlannelNodeConfig)
-
-	overlayAddressSpace := nodeConfig.OverlaySubnet.IPAddress + "/" + strconv.Itoa(int(nodeConfig.OverlaySubnet.PrefixLength))
-
-	containerPort, err := ovsctl.GetOVSPortNumber(ovs.HostVethName)
-	if err != nil {
-		return err
-	}
-	log.Printf("HERE Overlay Address Space: %v\n", overlayAddressSpace)
-	ovsctl.AddOverlayIPDnatRule(ovs.BridgeName, config.OverlayIP, ovs.ContainerIP, ovs.VlanID, containerPort)
-	ovsctl.AddOverlayIPSnatRule(ovs.BridgeName, containerPort, overlayAddressSpace, config.OverlayIP)
-	ovsctl.AddOverlayFakeArpReply(ovs.BridgeName, config.OverlayIP, ovs.ContainerMac)
-	return nil
-}
-
-func (fnp FlannelNephilaProvider) DeleteNetworkContainerRules(ovs NephilaOVSEndpoint, ncConfig interface{}) error {
-	/*
-		fncc := ncConfig.(NephilaNetworkContainerConfig)
-		config := fncc.Config.(FlannelNetworkContainerConfig)
-		nodeConfig := fncc.NodeConfig.(FlannelNodeConfig)
-
-		overlayAddressSpace := nodeConfig.OverlaySubnet.IPAddress + "/" + string(nodeConfig.OverlaySubnet.PrefixLength)
-
-		containerPort, err := ovsctl.GetOVSPortNumber(ovs.HostVethName)
-		hostPort, err := ovsctl.GetOVSPortNumber(ovs.HostPrimaryIfName)
-		if err != nil {
-			return err
-		}
-
-		ovsctl.DeleteOverlayIPDnatRule(ovs.BridgeName, hostPort, config.OverlayIP)
-		ovsctl.DeleteOverlayIPSnatRule(ovs.BridgeName, containerPort, overlayAddressSpace)
-		ovsctl.DeleteOverlayFakeArpReply(ovs.BridgeName, config.OverlayIP)
-	*/
-	return nil
-}
-
 func (fnp FlannelNephilaProvider) ConfigureNode(nodeConf NephilaNodeConfig, dncConf NephilaDNCConfig) (NephilaNodeConfig, error) {
 	var nodeConfig NephilaNodeConfig
 
@@ -106,4 +82,132 @@ func (fnp FlannelNephilaProvider) ConfigureNetworkContainerLink(link *netlink.VE
 	fNodeConf := ncConfig.NodeConfig.(FlannelNodeConfig)
 	link.LinkInfo.MTU = uint(fNodeConf.InterfaceMTU)
 	return nil
+}
+
+func SetFlannelKey(flannelDNCConfig FlannelDNCConfig) error {
+	pip := net.ParseIP(flannelDNCConfig.OverlaySubnet.IPAddress)
+	if pip == nil {
+		return fmt.Errorf("Failed to parse flannel overlay IP: %v", flannelDNCConfig.OverlaySubnet.IPAddress)
+	}
+	fetcd := flannelEtcdConfig{
+		Network: ip.IP4Net{
+			IP:        ip.FromIP(pip),
+			PrefixLen: uint(flannelDNCConfig.OverlaySubnet.PrefixLength),
+		},
+		SubnetLen: uint(flannelDNCConfig.PerNodePrefixLength),
+		Backend: flannelEtcdBackendConfig{
+			Type: vxlan,
+		},
+	}
+
+	b, err := json.Marshal(fetcd)
+	value := string(b)
+
+	cfg := client.Config{
+		Endpoints: []string{"http://127.0.0.1:2379"},
+		Transport: client.DefaultTransport,
+		// set timeout per request to fail fast when the target endpoint is unavailable
+		HeaderTimeoutPerRequest: time.Second,
+	}
+	c, err := client.New(cfg)
+	if err != nil {
+		return fmt.Errorf("Failed to create new etcd client with error %s", err)
+	}
+	kapi := client.NewKeysAPI(c)
+	// set "/foo" key with "bar" value
+	log.Printf("[Azure CNS Nephila: Flannel] Setting %s key in etcd with %s value.", flannelKeyPath, value)
+
+	resp, err := kapi.Set(context.Background(), flannelKeyPath, value, nil)
+	if err != nil {
+		return fmt.Errorf("Failed to set keys in etcd with error: %s", err)
+	}
+	log.Printf("[Azure CNS Nephila: Flannel] Set Flannel config in etcd with response %v.", resp)
+	return nil
+}
+
+func fileExists(filepath string) bool {
+	_, err := os.Stat(filepath)
+	return !os.IsNotExist(err)
+}
+
+// Handles retrieval of over
+func GetFlannelConfiguration() (*FlannelNodeConfig, error) {
+
+	subenv := "/var/run/flannel/subnet.env"
+
+	// run 5 attempts for flannel to write out file
+	for i := 0; i < 5; i++ {
+		if fileExists(subenv) {
+			break
+		}
+		log.Printf("Checking for subnet file (Attempt %v/5)\n", i+1)
+		time.Sleep(time.Second * 1)
+	}
+
+	fp, err := os.Open(subenv)
+
+	var flannel FlannelNodeConfig
+
+	if err != nil {
+		return nil, fmt.Errorf("Loading Flannel subnet file failed with error: %v", err)
+	} else {
+		log.Printf("Subnet file loaded\n")
+		defer fp.Close()
+
+		fenvs := make(map[string]string)
+		sr := bufio.NewScanner(fp)
+		for sr.Scan() {
+			ev := strings.Split(sr.Text(), "=")
+			fenvs[ev[0]] = ev[1]
+		}
+
+		// read allocatable space for the overlay
+		if v, exists := fenvs["FLANNEL_NETWORK"]; exists {
+			props := strings.Split(v, "/") // ex 169.254.22.1/24
+			flannel.OverlaySubnet.IPAddress = props[0]
+			prefix, err := strconv.ParseInt(props[1], 10, 32)
+			if err != nil {
+				return nil, fmt.Errorf("Flannel Network env failed to parse node subnet prefix.")
+			}
+			flannel.OverlaySubnet.PrefixLength = uint8(prefix)
+		} else {
+			return nil, fmt.Errorf("Flannel Subnet env not found.")
+		}
+
+		// read allocatable space for the subnet on node
+		if v, exists := fenvs["FLANNEL_SUBNET"]; exists {
+			props := strings.Split(v, "/") // ex 169.254.22.1/24
+			flannel.NodeSubnet.IPAddress = props[0]
+			prefix, err := strconv.ParseInt(props[1], 10, 32)
+			if err != nil {
+				return nil, fmt.Errorf("Flannel Network env failed to parse node subnet prefix.")
+			}
+			flannel.NodeSubnet.PrefixLength = uint8(prefix)
+		} else {
+			return nil, fmt.Errorf("Flannel Subnet env not found.")
+		}
+
+		if v, exists := fenvs["FLANNEL_MTU"]; exists {
+			mtu, err := strconv.ParseInt(v, 10, 32)
+			if err != nil {
+
+				return nil, errors.New("Flannel Network env failed to parse node MTU.")
+			}
+			flannel.InterfaceMTU = mtu
+		} else {
+			return nil, fmt.Errorf("Flannel MTU env not found.")
+		}
+
+		if v, exists := fenvs["FLANNEL_IPMASQ"]; exists {
+			ipmasq, err := strconv.ParseBool(v)
+			if err != nil {
+				return nil, fmt.Errorf("Error. Flannel Network env failed to parse IPMASQ.")
+			}
+			flannel.IPMASQ = ipmasq
+		} else {
+			return nil, fmt.Errorf("Error. Flannel IPMASQ env not found.")
+		}
+	}
+
+	return &flannel, nil
 }
