+++ network/bridge_endpointclient_linux.go
@@ -39,7 +39,7 @@ func NewLinuxBridgeEndpointClient(
 
 func (client *LinuxBridgeEndpointClient) AddEndpoints(epInfo *EndpointInfo) error {
 
-	if err := createEndpoint(client.hostVethName, client.containerVethName, epInfo.NephilaNCConfig); err != nil {
+	if err := createEndpoint(client.hostVethName, client.containerVethName, 0); err != nil {
 		return err
 	}
